refactor: move domains to usecases
